requirements.txt
openai-whisper
argostranslate
ffmpeg-python
pysrt
tqdm
TTS
torch
torchaudio
numpy
librosa
soundfile
flask-cors
webrtcvad
requests
pytest
pytest-timeout
transformers==4.35.2
